bullet flying trajectory changed

+++ src/main/java/org/game/towers/workers/Algorithms/MathAlgorithms/LinearAlgorithms.java
@@ -2,15 +2,15 @@ package org.game.towers.workers.Algorithms.MathAlgorithms;
 
 public class LinearAlgorithms {
 
-    public static int dependentYFromX(int x, int x1, int x2, int y1, int y2) {
+    public static double dependentYFromX(double x, int x1, int x2, int y1, int y2) {
         double aFactor = (x - x1) / (x2 - x1);
-        int y = (int)(y2 + ( aFactor * (y2 - y1) ));
+        double y = (int)(y2 + ( aFactor * (y2 - y1) ));
         return y;
     }
 
-    public static int dependentXFromY(int y, int x1, int x2, int y1, int y2) {
+    public static double dependentXFromY(double y, int x1, int x2, int y1, int y2) {
         double aFactor = (x2 - x1) / (y2 - y1);
-        int x = (int)(aFactor * (y - y2) + x1);
+        double x = (int)(aFactor * (y - y2) + x1);
         return x;
     }
